remsgr: add local structures for auxclk and regulator resources
Now, Host side is waiting for a name for auxclk and regulator resources.
So, in order to avoid doing changes in the exported APIs and all the
users the IpcResource module is translating the auxclk/regulator Id to
the needed name and passing that name to the Host resource manager.
For that reason we need to create a local structure that match the Host
side structure and keep the original structure to match exported APIs.

The old structures can be removed once all users of IpcResource are
migrated to the new changes.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>